backend/requirements.txt
# Web Framework and Server
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# API and Cloud Services
groq==0.4.2
python-dotenv==1.0.0
requests==2.31.0

# Database
pymongo==4.6.0

# Image Processing and Quality Checks
# Required for Blur, Brightness, and Contrast checks
opencv-python==4.8.1.78
numpy==1.26.2
# Required for basic image manipulation and Base64 encoding
pillow>=11.0.0

# PDF Processing
# Required for batch_processor.py to convert PDF pages to images
pymupdf==1.23.8
PyPDF2==3.0.1

# Data Parsing and Canonicalization
python-dateutil==2.8.2
dateparser==1.1.8
num2words==0.5.12
regex==2025.11.3